requirements.txt
asgiref==3.7.2
Django==3.2.21
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
install==1.3.5
packaging==23.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
